Convert some Gui elements to new format
