+++ src/ATGUI/Windows/configs.cpp
@@ -1,0 +1,106 @@
+#include "configs.h"
+
+bool Configs::showWindow = false;
+
+namespace ImGui
+{
+	static auto vector_getter1 = [](void* vec, int idx, const char** out_text)
+	{
+		auto& vector = *static_cast<std::vector<std::string>*>(vec);
+		if (idx < 0 || idx >= static_cast<int>(vector.size())) { return false; }
+		*out_text = vector.at(idx).c_str();
+		return true;
+	};
+
+	bool Combo1(const char* label, int* currIndex, std::vector<std::string>& values)
+	{
+		if (values.empty()) { return false; }
+		return Combo(label, currIndex, vector_getter1,
+					 static_cast<void*>(&values), values.size());
+	}
+
+	bool ListBox1(const char* label, int* currIndex, std::vector<std::string>& values, int height_in_items = -1)
+	{
+		if (values.empty()) { return false; }
+		return ListBox(label, currIndex, vector_getter1,
+					   static_cast<void*>(&values), values.size(), height_in_items);
+	}
+}
+
+
+void Configs::RenderWindow()
+{
+	if (!Configs::showWindow)
+		return;
+
+	ImGui::SetNextWindowSize(ImVec2(195, 260), ImGuiSetCond_Always);
+	if (ImGui::Begin("Configs", &Configs::showWindow, ImGuiWindowFlags_NoCollapse | ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_ShowBorders | ImGuiWindowFlags_NoResize))
+	{
+		static std::vector<std::string> configItems = GetConfigs();
+		static int configItemCurrent = -1;
+
+		if (ImGui::Button("Refresh"))
+			configItems = GetConfigs();
+
+		ImGui::SameLine();
+		if (ImGui::Button("Save"))
+		{
+			if (configItems.size() > 0 && (configItemCurrent >= 0 && configItemCurrent < (int) configItems.size()))
+			{
+				pstring path = GetConfigDirectory();
+				path << configItems[configItemCurrent] << "/config.json";
+
+				Settings::LoadDefaultsOrSave(path);
+			}
+		}
+
+		ImGui::SameLine();
+		if (ImGui::Button("Remove"))
+		{
+			if (configItems.size() > 0 && (configItemCurrent >= 0 && configItemCurrent < (int) configItems.size()))
+			{
+				pstring path = GetConfigDirectory();
+				path << configItems[configItemCurrent];
+
+				Settings::DeleteConfig(path);
+
+				configItems = GetConfigs();
+				configItemCurrent = -1;
+			}
+		}
+
+		static char buf[128] = "";
+		ImGui::PushItemWidth(138);
+		ImGui::InputText("", buf, IM_ARRAYSIZE(buf));
+		ImGui::PopItemWidth();
+
+		ImGui::SameLine();
+		if (ImGui::Button("Add") && strlen(buf) > 0)
+		{
+			pstring path = GetConfigDirectory();
+			path << buf;
+
+			if (!DoesFileExist(path.c_str()))
+			{
+				mkdir(path.c_str(), S_IRWXU | S_IRWXG | S_IROTH | S_IXOTH);
+				Settings::LoadDefaultsOrSave(path << "/config.json");
+
+				configItems = GetConfigs();
+				configItemCurrent = -1;
+			}
+		}
+
+		ImGui::PushItemWidth(178);
+		if (ImGui::ListBox1("", &configItemCurrent, configItems, 7))
+		{
+			pstring path = GetConfigDirectory();
+			path << configItems[configItemCurrent] << "/config.json";
+
+			Settings::LoadConfig(path);
+			UI::ReloadWeaponSettings();
+		}
+		ImGui::PopItemWidth();
+
+		ImGui::End();
+	}
+}
